Add game main logic

+++ src/utils.rs
@@ -1,6 +1,6 @@
 use rand::Rng;
-use std::collections::HashSet;
 use std::collections::BTreeMap;
+use std::collections::HashSet;
 
 pub fn next_with_wrap<T>(i: usize, collection: &Vec<T>) -> usize {
     return (i + 1) % collection.len();
@@ -10,7 +10,7 @@ pub fn previous_with_wrap<T>(i: usize, collection: &Vec<T>) -> usize {
     return (i + collection.len() - 1) % collection.len();
 }
 
-pub fn roll_dice() -> u32 {
+pub fn roll_dice() -> usize {
     let mut rng = rand::thread_rng();
     rng.gen_range(1..=6)
 }
